requirements.txt
requests
pandas
matplotlib
python-dateutil
python-dotenv
